feat(message-store): ✨ implement InMemoryMessageStore for message management
- Introduced a new InMemoryMessageStore class to manage message storage, retrieval, and removal.
- Implemented the IMessageStore interface to define the structure for message operations.
- Enhanced type safety and generics in message handling, allowing for flexible message criteria and payloads.
- Removed the previous implementation of IMessageStore and InMemoryMessageStore from the Message class, promoting a cleaner separation of concerns.
- These changes improve the maintainability and clarity of the messaging system, facilitating better message management.
